Improve ProcessTags() algorithm in daijisen.js.
Distinguish between tags and tag templates. The difference between a
tag and a tag template is that the tag template might contain some
parts, the presence of which in the final tag is completely optional.
With the new algorithm we permute each tag template to create
multiple real tags.

Results produced by the new algorithm improve 'magic seek' precision
considerably. Right before this change, there was a good chance that
the right article would be skipped because its tags had a tag
template inside.