Do character re-encoding inside hook functions.
With this change, the amount of complexity that's required to read
text from a dictionary has been substantially reduced. We used to
read text from the EPWING dictionary in two steps:

 1. Store text containing characters from different charsets into
    a temporary buffer, create a descriptor for each sequence of
    characters from the same charset.
 2. Encode text from the temporary buffer into UCS-2 using the
    descriptors that we've just created.

The two-step process was used to improve performance, because
calling iconv() for each character inside the hook functions was
slow.

With this change, encoding is performed inside the hook functions
using LUTs to map characters from foreign encodings to UCS-2.